+++ .gitignore
@@ -1,8 +1,3 @@
-Darts/Darts_Testvideo_8.mp4
-Darts/bkp
-Videos
-GUI/darts-arm-linux.zip
-GUI/darts-linux.zip
 *.pyc
 .idea
 .vscode
